Add ';' to HttpUrl.QUERY_COMPONENT_ENCODE_SET
This fixes a real-world issue I found that happens on GAE.

Also, I think having different encoded sets for different parts of the URL is kinda silly, and we should just encode everything the [RFC](http://www.faqs.org/rfcs/rfc3986.html) says we should encode, but that's better left for another patch. This is just a quick fix.